git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/cfo-branch@102276 138bc75d-0d04-0410-961f-82ee72b054a4

+++ libstdc++-v3/include/ext/pool_allocator.h
@@ -42,8 +42,8 @@
 
 /** @file ext/pool_allocator.h
  *  This file is a GNU extension to the Standard C++ Library.
- *  You should only include this header if you are using GCC 3 or later.
  */
+
 #ifndef _POOL_ALLOCATOR_H
 #define _POOL_ALLOCATOR_H 1
 
@@ -80,7 +80,7 @@ namespace __gnu_cxx
 
       enum { _S_align = 8 };
       enum { _S_max_bytes = 128 };
-      enum { _S_free_list_size = _S_max_bytes / _S_align };
+      enum { _S_free_list_size = (size_t)_S_max_bytes / (size_t)_S_align };
       
       union _Obj
       {
@@ -117,7 +117,7 @@ namespace __gnu_cxx
     };
 
 
-  /// @brief  Class __pool_alloc.
+  /// @brief  class __pool_alloc.
   template<typename _Tp>
     class __pool_alloc : private __pool_alloc_base
     {
